Create README.md
Created an application using React Native. 

Goal:
Building the application: 
Create an input box to take user input •
The input box should be associated with a button • 
When a user clicks the button, a list of tasks should be formed below • 
Each task much have a delete option •
Make use of React Hooks for state management.